Revert ezeka-file-link to return fully qualified links
This allows me to create notes in a specific Kasten, but links to existing notes
do not need to have the Kasten explicitly specified. In the background, links
should be fully qualified, but then Kasten should be stripped in display
functions like `ezeka-format-metadata` and `ezeka--format-link`, which should
always be used when inserting links.